Make setup dsl profile. Make reset settings to Default. Fixed update dslam port info. Fixed calculate value LastChange. Fixed calculate question result in BasicDialogs. Some refactoring.

+++ devices/dslamma5300.cpp
@@ -50,7 +50,7 @@ long DslamMa5300::snmpPortIndex(BoardType::Enum boardType, int slot,
 XdslPort::Ptr DslamMa5300::createPort(BoardType::Enum boardType, int boardIndex,
                                       int portIndex, QObject *parent)
 {
-    int snmpIndex = snmpPortIndex(boardType, boardIndex, portIndex);
+    long snmpIndex = snmpPortIndex(boardType, boardIndex, portIndex);
 
     switch (boardType) {
     case BoardType::AnnexA:
