+++ devices/dslamma5600.cpp
@@ -50,7 +50,7 @@ long DslamMa5600::snmpPortIndex(BoardType::Enum boardType, int slot,
 XdslPort::Ptr DslamMa5600::createPort(BoardType::Enum boardType, int boardIndex,
                                       int portIndex, QObject *parent)
 {
-    int snmpIndex = snmpPortIndex(boardType, boardIndex, portIndex);
+    long snmpIndex = snmpPortIndex(boardType, boardIndex, portIndex);
 
     switch (boardType) {
     case BoardType::AnnexA:
