+++ devices/dslammxa32.cpp
@@ -43,7 +43,7 @@ long DslamMxa32::snmpPortIndex(BoardType::Enum boardType, int slot,
 XdslPort::Ptr DslamMxa32::createPort(BoardType::Enum boardType, int boardIndex,
                                      int portIndex, QObject *parent)
 {
-    int snmpIndex = snmpPortIndex(boardType, boardIndex, portIndex);
+    long snmpIndex = snmpPortIndex(boardType, boardIndex, portIndex);
 
     switch (boardType) {
     case BoardType::AnnexA:
